agents/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bedrock_agentcore_example"
version = "0.1.0"
description = "Amazon Bedrock Agentcore Example for AI Builders Day on 2025/12/20"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.50",
    "strands-agents[otel]>=1.15.0",
    "strands-agents-tools>=0.2.14",
    "bedrock-agentcore>=1.1.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
